packages/rag/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Google Cloud / Vertex AI
google-cloud-aiplatform==1.71.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-auth==2.26.2
vertexai>=1.71.1

# Document processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0

# Vector operations
numpy==1.26.3

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

